// packages/ui-core/src/types.d.ts
export declare type Typography = {
  fontStyle: 'normal' | 'italic';
  fontWeight: 'normal' | 'bold';
  fontSize: number;
  fontColor: string;
  fontFamily?: string;
  lineHeight?: string;
};
export declare type Position = 'top' | 'bottom' | 'left' | 'right';
export declare type Alignment = 'left' | 'center' | 'right';
export declare type Layout = 'vertical' | 'horizontal';